feat(dtfs2-6966): update dtfs tests
