+++ dtfs-central-api/api-tests/helpers/create-deal.js
@@ -1,0 +1,5 @@
+import { generatePortalAuditDetails } from '@ukef/dtfs2-common/change-stream';
+
+export const createDeal = async ({ api, deal, user }) => {
+  return api.post({ deal, user, auditDetails: generatePortalAuditDetails(user._id) }).to('/v1/portal/deals');
+};
