+++ dtfs-central-api/api-tests/v1/facilities/facilities.api-test.js
@@ -5,6 +5,7 @@ const api = require('../../api')(app);
 const aDeal = require('../deal-builder');
 const { MOCK_DEAL } = require('../mocks/mock-data');
 const { MOCK_PORTAL_USER } = require('../../mocks/test-users/mock-portal-user');
+const { createDeal } = require('../../helpers/create-deal');
 
 const mockFacility = {
   type: 'Bond',
@@ -22,10 +23,6 @@ const newDeal = aDeal({
   },
 });
 
-const createDeal = async () => {
-  const { body } = await api.post({ deal: newDeal, user: MOCK_PORTAL_USER }).to('/v1/portal/deals');
-  return body;
-};
 describe('/v1/portal/facilities', () => {
   let dealId;
 
@@ -34,7 +31,7 @@ describe('/v1/portal/facilities', () => {
   });
 
   beforeEach(async () => {
-    const deal = await createDeal();
+    const { body: deal } = await createDeal({ api, deal: newDeal, user: MOCK_PORTAL_USER });
 
     dealId = deal._id;
     mockFacility.dealId = dealId;
