+++ dtfs-central-api/api-tests/v1/tfm/facilities/tfm-facility-get.api-test.js
@@ -7,6 +7,7 @@ const aDeal = require('../../deal-builder');
 const CONSTANTS = require('../../../../src/constants');
 const { MOCK_DEAL } = require('../../mocks/mock-data');
 const { MOCK_PORTAL_USER } = require('../../../mocks/test-users/mock-portal-user');
+const { createDeal } = require('../../../helpers/create-deal');
 
 const newFacility = {
   type: 'Bond',
@@ -24,11 +25,6 @@ const newDeal = aDeal({
   },
 });
 
-const createDeal = async () => {
-  const { body } = await api.post({ deal: newDeal, user: MOCK_PORTAL_USER }).to('/v1/portal/deals');
-  return body;
-};
-
 describe('/v1/tfm/facilities', () => {
   let dealId;
 
@@ -37,7 +33,7 @@ describe('/v1/tfm/facilities', () => {
   });
 
   beforeEach(async () => {
-    const deal = await createDeal();
+    const deal = await createDeal({ api, deal: newDeal, user: MOCK_PORTAL_USER });
     dealId = deal._id;
 
     newFacility.dealId = dealId;
