+++ dtfs-central-api/api-tests/v1/tfm/facilities/tfm-facility-put.api-test.js
@@ -9,6 +9,7 @@ const aDeal = require('../../deal-builder');
 const CONSTANTS = require('../../../../src/constants');
 const { MOCK_PORTAL_USER } = require('../../../mocks/test-users/mock-portal-user');
 const { MOCK_TFM_USER } = require('../../../mocks/test-users/mock-tfm-user');
+const { createDeal } = require('../../../helpers/create-deal');
 
 const newFacility = {
   type: 'Bond',
@@ -29,10 +30,6 @@ const newDeal = aDeal({
   },
 });
 
-const createDeal = async () => {
-  const { body } = await api.post({ deal: newDeal, user: MOCK_PORTAL_USER }).to('/v1/portal/deals');
-  return body;
-};
 describe('/v1/tfm/facilities', () => {
   let dealId;
 
@@ -41,7 +38,7 @@ describe('/v1/tfm/facilities', () => {
   });
 
   beforeEach(async () => {
-    const deal = await createDeal();
+    const deal = await createDeal({ api, deal: newDeal, user: MOCK_PORTAL_USER });
 
     dealId = deal._id;
     newFacility.dealId = dealId;
